requirements.txt
gradio==4.44.1
huggingface_hub<0.23.0
litellm>=1.0.0
python-dotenv>=1.0.0
